python module: Use sys_root's Python sysconfigdata to get EXT_SUFFIX/SO
Until now, if the build machine's Python was used for cross, the wrong
EXT_SUFFIX/SO was used, resulting in native extensions being built and
installed with the wrong filename.

To fix this, we load the sys_root sysconfigdata from its stdlib
directory. We assume that the same Python version as the build machine
exists within the sys_root with the same stdlib directory, save for the
prefix. If not, we fall back to the build machine, as we did before.

If the machine file points python at a Python within sys_root then this
should still work. You should probably only run a sys_root's binary
using some kind of wrapper, but either way, sysconfig will still return
the correct values.

We only read EXT_SUFFIX/SO from the sys_root because paths are expected
to be the same, save for the prefix, and other info such as the platform
and limited API suffix cannot be determined this way.

We could potentially guess the limited API suffix from the platform, as
there are only a small number of possible values, but this can be done
separately.

Closes: https://github.com/mesonbuild/meson/issues/7049